cart page implementation completed

+++ src/app/api/cart/removeFromCart/route.ts
@@ -3,6 +3,8 @@ import { CartModel } from "@/models/carts.model";
 import { Types } from "mongoose";
 export async function POST(req: NextRequest) {
   const { productId } = await req.json();
+  const _id = req.cookies.get("_id")?.value;
+  const userId = new Object(_id)
   if (!productId) {
     return NextResponse.json(
       { message: "Missing productId in request body", status: 400, success: false },
@@ -11,7 +13,7 @@ export async function POST(req: NextRequest) {
   }
   try {
     const productObjectId = new Object(productId);
-    const cartItem = await CartModel.findOneAndDelete({ productId: productObjectId });
+    const cartItem = await CartModel.findOneAndDelete({ userId,productId: productObjectId });
     if (!cartItem) {
       return NextResponse.json(
         { message: "Cart item not found", status: 404, success: false },
